♻️ Move Hit card to its own file
