+++ components/Search/index.ts
@@ -1,2 +1,3 @@
 export { SearchResultStats } from "./SearchResultStats"
 export { Filter } from "./Filter"
+export { Hit } from "./Hit"
